Replaced ssize_t for int to avoid potetial underflow

+++ src/remoted/secure.c
@@ -74,7 +74,7 @@ void HandleSecure()
     int sock_client;
     int n_events = 0;
     char buffer[OS_MAXSTR + 1];
-    ssize_t recv_b;
+    int recv_b;
     struct sockaddr_in peer_info;
     memset(&peer_info, 0, sizeof(struct sockaddr_in));
     wnotify_t * notify = NULL;
